# [libc++] Optimize vector growing of trivially relocatable types