@ldionne 2024-02-01
Let's fix the last nit and land this. Thanks for this patch, it's a really important optimization.