@kstoimenov 2024-02-02
This was reverted in 67eee4a029797c09129889c3655416d1be487cfe due to sanitizer build bot breaking.