@philnik777 2024-02-02
@kstoimenov Could you maybe give me some more information? "My bot broke" is really not helpful, and I'm not at all familiar with any of the non-libc++ code that is shown in the traces.